trending-analyzer/requirements.txt
pyspark>=3.4.0 # Use a version compatible with your Spark cluster
kafka-python>=2.0.0
redis>=4.0.0
python-dotenv>=0.20.0 # For loading .env files 